refactor: common 컴포넌트 import 결합
