+++ src/components/@common/index.js
@@ -1,0 +1,4 @@
+export { default as Button } from 'components/@common/Button';
+export { default as FieldSet } from 'components/@common/FieldSet';
+export { default as Header } from 'components/@common/Header';
+export { default as TextField } from 'components/@common/TextField';
